+++ Sandboxie/core/svc/serviceserver.h
@@ -54,10 +54,10 @@ public:
 
 private:
 
-    bool CanCallerDoElevation(
+    static bool CanCallerDoElevation(
             HANDLE idProcess, const WCHAR *ServiceName, ULONG *pSessionId);
 
-	bool CanAccessSCM(HANDLE idProcess);
+    static bool CanAccessSCM(HANDLE idProcess);
 
     static void ReportError2218(HANDLE idProcess, ULONG errlvl);
 
@@ -68,7 +68,7 @@ private:
     MSG_HEADER *RunHandler(MSG_HEADER *msg, HANDLE idProcess);
 
     ULONG RunHandler2(
-        HANDLE idProcess, ULONG idSession,
+        HANDLE idProcess, ULONG idSession, ULONG type,
         const WCHAR *devmap, const WCHAR *svcname, const WCHAR *path);
 	
     MSG_HEADER *UacHandler(